requirements.txt
flask
flask-cors
torch>=1.3.1
numpy==1.17.0
scikit-learn
matplotlib
imageio
Pillow
imgaug
tqdm
opencv-contrib-python==4.5.1.48